feature delete category
